Implement walking of directories
Recurse into directories using BurntSushi's walkdir crate. This means
the following is now possible:

    unicop src/

Also set the default to '.' (i.e. recursively checking the current
directory).

I've read in the documentation that tree-sitter actually has a
language-detection framework that grammars can plug into, but
unfortunately, it doesn't seem to be exposed in the library. It might
still be a good idea to adopt it, as it will make it easier to add new
languages. This commit contains a stripped-down version that only look
at file extensions (tree-sitter's language detection also uses regexps,
but none of the current grammars define any.)