Only print about parsing errors in verbose mode
The rationale is that parsing errors are normally not much of a security
issue. Code that fails to parse fails to map each code point to the
correct code type. This means that the code that can't be parsed
is treated as "other" code rather than the type specific code. And the
rules for "other" is usually more strict than the type specific rules.
It is for example common (and the default) to be more relaxed in
comments and string literals, than in code in general.